src/video_8/requirements.txt
kuzu==0.7.0
python-dotenv==1.0.1
llama-index==0.12.2
llama-index-embeddings-openai~=0.3.0
llama-index-llms-openai~=0.3.0
llama-index-graph-stores-kuzu~=0.6.0
